requirements.txt
tweepy
python-dotenv
pandas
numpy
matplotlib
wordcloud
textblob
vaderSentiment
nltk
scikit-learn
jupyter
streamlit